Fix issue with file extension parsing and yaml file rendering.
This changes fixes two bugs introduced in 1ca427e33. The first is
that variable files extensions are not being matched due to a
change in the extension discovery which now means the `.` is
included whereas previously it was not.

The second fixes an issue where templates using a yaml/yml file
extension were not being rendered.

Closes #4
Closes #5