Nomad jobs of type batch do not produce evaluations.
Previously Levant would register the job and then check the return
for the eval_ID and thus continue with the deployment checking.
Nomad batch jobs do not produce an evaluation upon registration
and therefore would cause an error within Levant which was not
handled in the most appropriate manner. This change checks the job
type and eval return to properly catch batch job registrations and
log correct information.

In the future Levant should be updated to perform a basic running
check on batch jobs to provide further deployment feedback.

Closes #50