Levant can now track Nomad auto-revert of a failed deployment.
As of Nomad 0.6 deployments supported auto-revert; whereby a job
would be reverted to its previous stable configuration if a
deployment was marked as failed. This change introduces the
ability for Levant to track the auto-revert deployment and log
realtime feedback. Levant will also print in CAPS if either a
revert fails or a deployment failed and is not configured to
revert so that users can make adjustments to their job in future
to avoid potential outage situations.

Closes #47