Update Vendored version of Nomad to 0.7.1.
The major release 0.7 of Nomad introduced ALCs amongst a raft
of other features. Therefore Levant requires updating to allow
users who are running the latest version of Nomad and using ACLs
to be able to use Levant.

The README.md has also been updated to provide a quick section on
what environment variables to be used by Levant and the Nomad
client to configure certain parameters.

Closes #75