Fix issue where jobs without specified count caused panic.
This issue was caused by the incorrect assumption that jobs with
a count not set or set to 0 can not result in a deployment. This
was incorrect as an undefined count will result in a default count
of 1; or that jobs set to count=0 can be run and deployed to stage
the job or disable it.

Closes #87