Addition of scale-out and scale-in commands.
This commit is the initial commit of code to allow users to
manually scale a job in or out manually. Scaling can be done either
by a count or a percentage value with option to pass a taskgroup
name to target. If the task group is not passed, all TGs within the
job will be scaled as dictated.

Closes #124