Addition of template functions allow Consul and Time lookups.
This commit introduces template functions which allow lookups
against Consul as well as generic time lookups which are run over
the template during execution. A consul client has been added with
config flag in both the render and deploy commands as these may
need to connect to Consul for variable lookups.

Closes #58