Use missingkey=zero rather than error in template rendering.
When performing complex template rendering, it is helpful to be
able to perform if statements based on the existence of a passed
variable like `if .env`. The template runs fine if the variable
is passed, but if it isn't the execution will stop because of the
missingkey setting. This change proposes setting the missingkey
value to zero rather than error. This means that rather than error
out, the operation returns the zero value for the map type's
element and the render execution can continue.