Add force flag to trigger a deployment even without changes
In some cases it is necessary to deploy a version that was already
deployed. For example, when something unexpected happened
on the Nomad side and not all allocations come up.
To mitigate that, the `force` flag triggers a deployment without
running `plan` before and just blindly forwards the command
to Nomad.