Use deployment watcher if there's a deploy ID
Previously levant would check the template contained an update stanza at
the job level, before watching for a service job deployment to complete.

With this change, we'll always wait for a deployment to complete. This
allows the Nomad server to be the one to determine if the template has a
"deployment" rather than levant.