template: move jobspec parsing into Levant to avoid dep hell.
The Nomad module setup is not designed nor intended to have any
pkg apart from api imported. This caused problems when using the
jobspec package. The Nomad Terraform provider has exactly this
problem and has its own version of the Parse functionality. I
attempted to import this to avoid more duplication but ran into
https://github.com/hashicorp/terraform-provider-nomad/issues/150.

This is therefore a temp solution so the project can be easily
used and released.