template: remove use of Terraform config for variable parsing.
Levant used an old version of Terraform to perform the .tf file
parsing. The functionality has been removed from TF and has been
causing go get problems. It is also the incorrect action to
perform, treating the file like a module and picking the variables
out.

This change moves to using pure HCL to parse the TF var file which
removes the dep and also improves the code. It also opens the door
to adding new, better variables types such as tfvars style which
us probably better suited to the Levant usecase.