Signal and fail deployment when no eligible nodes are found
I run into this problem by writing wrong datacenter name in the nomadfile.
There where no nodes in that misspelled datacenter but the deployment
run forever waiting for the node to appear without indication of the
problem.
Nomad UI is showing the warning that there are no nodes eligible for
evaluation. But there is no such warning in levant output. So I added
the same warning as in UI.
Instead of waiting for some node to appear we are now failing deployment.