cli: allow nested variables to be passed via var CLI flag.
This change updates the way the var CLI flag is processed to allow
the use of nested variables in the form of nested.var=val. Until
now either side of the delimiter was evaluated simply as a string.

All tests have been updated and accomodate the change aongside the
acceptance tests.