docs: clarify source for official release binaries
The Download & Install instructions are accurate but the order of
those instructions makes it easy for someone to start by downloading
older versions via `curl` and then miss the further instructions for
downloading current versions and official builds.

Move the HashiCorp releases option to be first, move the older version
releases to be last, and clean up the `go install` invocation.